Use uintmax_t for the setrlimit type
This avoids errors when calling it with RLIM_INFINITY.

+++ libminijail.h
@@ -128,7 +128,7 @@ void minijail_inherit_usergroups(struct minijail *j);
 int minijail_use_alt_syscall(struct minijail *j, const char *table);
 
 /* Sets the given runtime limit. See getrlimit(2). */
-int minijail_rlimit(struct minijail *j, int type, uint32_t cur, uint32_t max);
+int minijail_rlimit(struct minijail *j, int type, uintmax_t cur, uintmax_t max);
 
 /*
  * Adds the jailed process to the cgroup given by |path|.  |path| should be the
